clean order history
